requirements.txt
Django==3.2
django-environ==0.4.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
uWSGI
Pillow==8.3.1
requests==2.26.0
environ
psycopg2
django-environ
celery==5.1.2
django_redis
